Fix locking issues in private proxy (hopefully) and improve announcements from private to public proxy by bundling them.

+++ jsonrpc/context.go
@@ -63,7 +63,7 @@ func (c *Context) Result(data interface{}) *Response {
 	}
 }
 
-func (c *Context) Error(code int, message string, data interface{}) *Response {
+func (c *Context) Error(code int64, message string, data interface{}) *Response {
 	return &Response{
 		Error: &Error{
 			Code:    code,
